+++ ast/location_error.mli
@@ -6,3 +6,5 @@ val message : t -> string
 val set_message : t -> string -> t
 val make : loc:Location.t -> string -> sub:(Location.t * string) list -> t
 val to_extension : t -> Import.Parsetree.extension
+val raise : t -> 'a
+val update_loc : t -> Location.t -> t